refactored code to remove console logs and fixed bugs
