+++ src/app/FrontendTypes.ts
@@ -384,3 +384,16 @@ export interface Snapshots {
 export interface ErrorContainerProps {
   port: chrome.runtime.Port | null;
 }
+
+export interface AxContainer {
+  axSnapshots: [];
+  snapshot: {
+    name?: string;
+    componentData?: object;
+    state?: string | object;
+    stateSnaphot?: object;
+    children?: any[];
+  };
+  snapshots: [];
+  currLocation: object;
+}
